backend/foodgram/requirements.txt
django==3.2.15
flake8==5.0.4
python-dotenv==0.21.0
djoser==2.1.0
djangorestframework==3.14.0
django-filter==22.1
Pillow==9.2.0
psycopg2==2.9.4
drf-yasg==1.21.4
